track_record/requirements.txt
# portfolio_analyzer/requirements.txt
streamlit
pandas
requests
blockfrost-python
python-dateutil